jsonrpc::client
Struct Endpoint

pub struct Endpoint<I: AsyncRead, O: AsyncWrite> { /* private fields */ }

An endpoint for RPC calls.

Implementations

impl<I: AsyncRead + Unpin, O: AsyncWrite + Unpin> Endpoint<I, O>

pub fn new(input: I, output: O) -> Self

Create an enpdoint that reads from input and writes to output

pub async fn call<A, R>( &mut self, method: impl AsRef<str>, arg: A, ) -> Result<R, JsonRpcError>
where A: Serialize, R: DeserializeOwned,

Call the RPC method method with argument arg; decode and return the output

pub async fn batch_call<A: Serialize, R: DeserializeOwned>( &mut self, method: impl AsRef<str>, args: impl IntoIterator<Item = A>, ) -> Result<impl Iterator<Item = R>, JsonRpcError>

Call the method method once for each argument in args using a JSON RPC batch request and await all of the responses. Return the results of the calls in the same order as args.